Rename cache.py to persistence.py
This will make it less confusing when we introduce a new user-facing
Cache class.
